% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-sparsebnData.R
\name{as.data.frame.sparsebnData}
\alias{as.data.frame.sparsebnData}
\title{Convert a sparsebnData object back to a data.frame}
\usage{
\method{as.data.frame}{sparsebnData}(x)
}
\description{
Convert a sparsebnData object back to a data.frame
}

