% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-edgeList.R
\name{as.edgeList}
\alias{as.edgeList}
\title{as.edgeList}
\usage{
as.edgeList(x)
}
\arguments{
\item{x}{A compatible \code{R} object.}
}
\value{
\code{\link{edgeList}}
}
\description{
Methods for coercing other \code{R} objects to \code{\link{edgeList}} objects.
}
