% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-sparse.R
\name{as.sparse}
\alias{as.sparse}
\title{as.sparse}
\usage{
as.sparse(x, index = "R", ...)
}
\arguments{
\item{x}{A compatible \code{R} object.}

\item{index}{\code{"R"} or \code{"C"}, depending on whether to use R- or C-style indexing.}

\item{...}{other parameters.}
}
\value{
\code{\link{sparse}}
}
\description{
Methods for coercing other \code{R} objects to \code{\link{sparse}} objects.
}
