% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-generics.R
\name{coerce_discrete}
\alias{coerce_discrete}
\title{Recode discrete data}
\usage{
coerce_discrete(x)
}
\arguments{
\item{x}{an R object to coerce.}
}
\description{
Recodes discrete data so that the levels correspond to \code{0...n-1}, where \code{n}
is the total number of levels in a discrete factor.
}
\details{
Assumes data is unordered. Ordered factors are not supported at this time.
}
\examples{
x <- 1:5
coerce_discrete(x) # output: 0 1 2 3 4

x <- c("high", "normal", "high", "low")
coerce_discrete(x) # output: 0 2 0 1

}

