% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-sparsebnData.R
\name{count.interventions}
\alias{count.interventions}
\title{Count the number of rows under intervention}
\usage{
count.interventions(data)
}
\arguments{
\item{data}{a \code{\link{sparsebnData}} object.}
}
\description{
Returns the number of rows with at least one intervention
}
