% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-sparsebnData.R
\name{count.levels}
\alias{count.levels}
\title{Count the number of levels per variable}
\usage{
count.levels(data)
}
\arguments{
\item{data}{a \code{\link{sparsebnData}} object.}
}
\description{
Returns the number of levels per variable as an ordered vector.
}
