% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-edgeList.R
\name{degrees}
\alias{degrees}
\title{Degree distribution of a graph}
\usage{
degrees(x)
}
\arguments{
\item{x}{an \code{edgeList} object}
}
\description{
Returns a \code{data.frame} with summary statistics on the total degree,
in-degree, and out-degree of each node in the network.
}
