% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-edgeList.R
\docType{class}
\name{edgeList-class}
\alias{edgeList-class}
\title{edgeList class}
\description{
Convenience wrapper class for a (column-major) edge list. Each component of the list
corresponds to a node, and each component is an integer vector whose components are the parents
of this node in the graph. Only intended for internal use.
}
\details{
Also inherits from \code{\link{list}}.
}
\section{Methods}{

\code{\link{get.adjacency.matrix}},
\code{\link{num.nodes}}, \code{\link{num.edges}}
}

