% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-generics.R
\name{estimate.covariance.matrix}
\alias{estimate.concentration.matrix}
\alias{estimate.covariance.matrix}
\title{estimate.covariance.matrix}
\usage{
estimate.covariance.matrix(fit, data, ...)

estimate.concentration.matrix(x, ...)
}
\value{
\code{matrix}
}
\description{
Estimates the covariance (or concentration) matrix implied by the associated graph object.
}

