% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsebnUtils-inference.R
\name{fit_dag}
\alias{fit_dag}
\title{Inference in Bayesian networks}
\usage{
fit_dag(parents, dat, call = "lm.fit", ...)
}
\arguments{
\item{parents}{\code{\link{edgeList}} object.}

\item{dat}{Data.}

\item{call}{Either \code{"lm.fit"} or \code{"glm.fit"}.}

\item{...}{If \code{call = "glm.fit"}, specify \code{family} here. Also allows for other parameters to \code{lm.fit} and \code{glm.fit}.}
}
\description{
Basic computing engine called by \code{\link{estimate.parameters}} for fitting parameters
in a Bayesian network. Should not be used directly unless by experienced users.
}
\details{
Can call either \code{\link{lm.fit}} or \code{\link{glm.fit}}, with any choice of family.
}

