% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsebnUtils-inference.R
\name{fit_multinom_dag}
\alias{fit_multinom_dag}
\title{A function to do inference in Bayesian network.}
\usage{
fit_multinom_dag(parents, n_levels, dat)
}
\arguments{
\item{parents}{An edgeList object.}

\item{n_levels}{A vector indicating number of levels for each variable}

\item{dat}{Data, a dataframe or matrix}
}
\value{
List, of length the number of nodes. ith entry of the list is another list that contains parents of node i and the intercept coefficient for the ith node. For each parent of node i, it is a list of index of parent and the coefficient matrix of influence of parent has on node i.
}
\description{
A function to do inference in Bayesian network.
}

