% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsebnUtils-lambdas.R
\name{generate.lambdas}
\alias{generate.lambdas}
\title{generate.lambdas}
\usage{
generate.lambdas(lambda.max, lambdas.ratio = 0.001, lambdas.length = 50,
  scale = "linear")
}
\arguments{
\item{lambda.max}{Maximum value of lambda; in terms of the algorithm this is the initial value
of the regularization parameter in the solution path.}

\item{lambdas.ratio}{Ratio between the maximum lambda value and the minimum lambda value in the solution
path.}

\item{lambdas.length}{Number of values to include.}

\item{scale}{Which scale to use: Either \code{"linear"} or \code{"log"}.}
}
\description{
Convenience function for creating a grid of lambdas.
}
\details{
See Section 5.3 of \href{http://jmlr.org/papers/v16/aragam15a.html}{Aragam and Zhou (2015)}
for a discussion of regularization paths (also, solution paths).
}
