% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-edgeList.R, R/s3-generics.R,
%   R/s3-sparsebnFit.R, R/s3-sparsebnPath.R
\name{get.adjacency.matrix.edgeList}
\alias{get.adjacency.matrix.edgeList}
\alias{get.adjacency.matrix}
\alias{get.adjacency.matrix.sparsebnFit}
\alias{get.adjacency.matrix.sparsebnPath}
\title{get.adjacency.matrix}
\usage{
\method{get.adjacency.matrix}{edgeList}(x)

get.adjacency.matrix(x)

\method{get.adjacency.matrix}{sparsebnFit}(x)

\method{get.adjacency.matrix}{sparsebnPath}(x)
}
\arguments{
\item{x}{any \code{R} object.}
}
\value{
\code{matrix}
}
\description{
Extracts the adjacency matrix of the associated graph object.
}
\section{Methods (by class)}{
\itemize{
\item \code{edgeList}: Convert internal \code{edgeList} representation to an adjacency matrix

\item \code{sparsebnFit}: Retrieves \code{edges} slot and converts to an adjacency matrix

\item \code{sparsebnPath}: Retrieves all \code{edges} slots in the solution path, converts to an adjacency matrix, and returns as a list
}}

