% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-generics.R
\name{get.covariance.matrix}
\alias{get.concentration.matrix}
\alias{get.covariance.matrix}
\title{get.covariance.matrix}
\usage{
get.covariance.matrix(x, ...)

get.concentration.matrix(x, ...)
}
\value{
\code{matrix}
}
\description{
Computes the implied covariance (or concentration) matrix of the associated graph object.
}

