% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-generics.R
\name{get.covariance}
\alias{get.covariance}
\alias{get.precision}
\title{get.covariance}
\usage{
get.covariance(x, ...)

get.precision(x, ...)
}
\value{
\code{matrix}
}
\description{
Computes the implied covariance (or concentration) matrix of the associated graph object.
}

