% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-generics.R, R/s3-sparsebnPath.R
\name{get.lambdas}
\alias{get.lambdas}
\alias{get.lambdas.sparsebnPath}
\title{get.lambdas}
\usage{
get.lambdas(x)

\method{get.lambdas}{sparsebnPath}(x)
}
\arguments{
\item{x}{a \code{\link{sparsebnPath}} object.}
}
\value{
Vector of \code{numeric} lambda values in fitted object.
}
\description{
Extracts the lambda values from a \code{\link{sparsebnPath}} object.

Extract regularization path from solution path
}
\section{Methods (by class)}{
\itemize{
\item \code{sparsebnPath}: Returns a vector of lambda values defining the solution path of a \code{\link{sparsebnPath}} object.
}}

