% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-generics.R, R/s3-sparsebnFit.R,
%   R/s3-sparsebnPath.R
\name{get.nodes}
\alias{get.nodes}
\alias{get.nodes.sparsebnFit}
\alias{get.nodes.sparsebnPath}
\title{get.nodes}
\usage{
get.nodes(x)

\method{get.nodes}{sparsebnFit}(x)

\method{get.nodes}{sparsebnPath}(x)
}
\arguments{
\item{x}{a \code{\link{sparsebnFit}} or \code{\link{sparsebnPath}} object.}
}
\value{
Vector of \code{character} names.
}
\description{
Returns the node names associated with a fitted object.
}
\section{Methods (by class)}{
\itemize{
\item \code{sparsebnFit}: Returns the node names from a \code{\link{sparsebnFit}} object.

\item \code{sparsebnPath}: Returns the node names from a \code{\link{sparsebnPath}} object.
}}

