% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-edgeList.R, R/s3-generics.R
\name{is.zero.edgeList}
\alias{is.zero}
\alias{is.zero.edgeList}
\title{is.zero}
\usage{
\method{is.zero}{edgeList}(edgeL)

is.zero(x)
}
\description{
Determines whether or not the object is the same as the null or zero object from its class.
}
\section{Methods (by class)}{
\itemize{
\item \code{edgeList}: Determines whether or not the object represents a null graph with no edges.
}}

