% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-generics.R
\name{lambda.grid}
\alias{lambda.grid}
\title{lambda.grid}
\usage{
lambda.grid(x)
}
\arguments{
\item{x}{a \code{\link{sparsebnPath}} object.}
}
\value{
Vector of \code{numeric} lambda values in fitted object.
}
\description{
Extracts the lambda values from a \code{\link{sparsebnPath}} object.
}

