% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-edgeList.R, R/s3-generics.R, R/s3-sparsebnFit.R, R/s3-sparsebnPath.R
\name{num.nodes.edgeList}
\alias{num.nodes}
\alias{num.nodes.edgeList}
\alias{num.nodes.sparsebnFit}
\alias{num.nodes.sparsebnPath}
\title{num.nodes}
\usage{
\method{num.nodes}{edgeList}(edgeL)

num.nodes(x)

\method{num.nodes}{sparsebnFit}(fit)

\method{num.nodes}{sparsebnPath}(path)
}
\description{
Extracts the number of nodes of the associated graph object.
}
\section{Methods (by class)}{
\itemize{
\item \code{edgeList}: Extracts the number of nodes of \link{edgeList} object.

\item \code{sparsebnFit}: Extracts the number of nodes of \link{sparsebnFit} object.

\item \code{sparsebnPath}: Extracts the number of nodes of \link{sparsebnPath} object.
}}

