% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-generics.R, R/s3-sparsebnData.R, R/s3-sparsebnFit.R, R/s3-sparsebnPath.R
\name{num.samples}
\alias{num.samples}
\alias{num.samples.sparsebnData}
\alias{num.samples.sparsebnFit}
\alias{num.samples.sparsebnPath}
\title{num.samples}
\usage{
num.samples(x)

\method{num.samples}{sparsebnData}(sbd)

\method{num.samples}{sparsebnFit}(cf)

\method{num.samples}{sparsebnPath}(cp)
}
\description{
Extracts the number of samples of the associated object.
}
\section{Methods (by class)}{
\itemize{
\item \code{sparsebnData}: Extracts the number of samples of \link{sparsebnData} object.

\item \code{sparsebnFit}: Extracts the number of samples of \link{sparsebnFit} object.

\item \code{sparsebnPath}: Extracts the number of samples of \link{sparsebnPath} object.
}}

