% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-edgeList.R
\name{permute.nodes}
\alias{permute.nodes}
\title{Permute the order of nodes in a graph}
\usage{
permute.nodes(x)
}
\arguments{
\item{x}{Graph as \code{\link{edgeList}} object.}
}
\value{
Permuted graph as \code{\link{edgeList}} object.
}
\description{
Randomize the order of the nodes in a graph.
}
\details{
Useful for obfuscating the
topological sort in a DAG, which is often the default output of methods
that generate a random DAG. Output is graph isomorphic to input
}

