% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-sparsebnPath.R
\name{print.sparsebnPath}
\alias{print.sparsebnPath}
\title{print.sparsebnPath}
\usage{
\method{print}{sparsebnPath}(x, verbose = FALSE)
}
\arguments{
\item{verbose}{If \code{TRUE}, then each estimate in the solution path is printed separately. Do not use for
large graphs or large solution paths. (default = \code{FALSE})}
}
\description{
Prints the contents of a \code{\link{sparsebnPath}} object neatly.
}

