% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsebnUtils-generate.R
\name{random.graph}
\alias{random.graph}
\title{Generate random DAGs}
\usage{
random.graph(nnode, nedge, acyclic = TRUE, loops = FALSE)
}
\arguments{
\item{nnode}{Number of nodes in the graph.}

\item{nedge}{Number of edges in the graph.}

\item{acyclic}{If \code{TRUE}, output will be an acyclic graph.}

\item{loops}{If \code{TRUE}, output may include self-loops.}
}
\value{
An \code{\link[sparsebnUtils]{edgeList}} object containing a list of parents for each node.
}
\description{
Generate a random graph with fixed number of edges.
}
