% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsebnUtils-generate.R
\name{random.spd}
\alias{random.spd}
\title{Generate a random positive definite matrix}
\usage{
random.spd(nnode, eigenvalues = NULL, num.ortho = 10)
}
\arguments{
\item{nnode}{Number of nodes in the matrix}

\item{eigenvalues}{Vector of eigenvalues desired in output. If this
has fewer than nnode values, the remainder are filled in as zero.}

\item{num.ortho}{Number of random Householder reflections to compose}
}
\description{
Generate a random positive definite matrix
}
