% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsebnUtils-options.R
\name{setPlotPackage}
\alias{setPlotPackage}
\alias{getPlotPackage}
\title{Change default plotting mechanism}
\usage{
setPlotPackage(pkg)

getPlotPackage()
}
\arguments{
\item{pkg}{The desired package; default value is \code{igraph}.}
}
\description{
Changes the default plotting mechanism used by \code{sparsebn} to plot output and fitted objects.
}
\details{
For plotting, \code{sparsebn} can use one of three packages: \code{graph} (see also \code{Rgraphviz}),
\link[igraph]{igraph} (see \link[igraph]{plot.igraph}), and \link[network]{network-package} (see \link[network]{plot.network}).
Note that plotting requires that (at least one of) these extra packages are installed.
}
\section{Functions}{
\itemize{
\item \code{getPlotPackage}: Returns the current choice of plotting mechanism
}}

\seealso{
\code{\link{setGraphPackage}}, \code{\link{getGraphPackage}}
}
