% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-generics.R, R/s3-sparse.R
\docType{class}
\name{sparse}
\alias{as.sparse}
\alias{is.sparse}
\alias{sparse}
\title{sparse class}
\usage{
sparse(x, ...)

as.sparse(x, ...)

is.sparse(x)
}
\arguments{
\item{x}{Various \code{R} objects.}

\item{...}{(optional) additional arguments.}
}
\description{
Low-level representation of sparse matrices
}
\details{
An alternative data structure for storing sparse matrices in R using the (row, column, value)
format. Internally it is stored as a list with three components, each vectors, that contain
the rows / columns / values of the nonzero elements.
}

