% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-edgeList.R, R/s3-generics.R, R/s3-sparsebnPath.R
\name{is.edgeList}
\alias{is.edgeList}
\alias{is.sparsebnPath}
\alias{sparsebn-classes}
\title{Internal generics for fitted objects and data}
\usage{
is.edgeList(x)

is.sparsebnPath(x)
}
\description{
Internal methods provided by \code{sparsebnUtils} for development.
}

