% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-sparsebnData.R
\docType{class}
\name{sparsebnData-class}
\alias{sparsebnData-class}
\title{sparsebnData class}
\description{
Convenience wrapper class for observational and experimental data.
}
\details{
Also inherits from \code{\link{list}}.
}
\section{Methods}{

\code{\link{first.method.name}}
\code{\link{second.method.name}}
}

