% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-generics.R, R/s3-sparsebnFit.R
\name{to_edgeList}
\alias{to_edgeList}
\alias{to_edgeList.sparsebnFit}
\title{Conversion to edgeList object}
\usage{
to_edgeList(x)

\method{to_edgeList}{sparsebnFit}(x)
}
\arguments{
\item{x}{An object of type \code{\link{sparsebnPath}}, \code{\link{sparsebnFit}}, \code{\link[graph]{graphNEL-class}},
\code{\link[igraph]{igraph}}, or \code{\link[network]{network}}.}
}
\description{
\code{to_edgeList} converts an object to an \code{\link{edgeList}} object. Works on both fitted
objects and graphs themselves. In the first case, every underlying 'edges' component is converted to
\code{\link{edgeList}}. In the second, the conversion applies directly to the object.
}
\section{Methods (by class)}{
\itemize{
\item \code{sparsebnFit}: description
}}

