% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-compatibility.R
\name{to_graphNEL}
\alias{to_graphNEL}
\title{Conversion to graphNEL objects}
\usage{
to_graphNEL(x)
}
\arguments{
\item{x}{An object of type \code{\link{edgeList}}, \code{\link{sparsebnFit}}, or \code{\link{sparsebnPath}}.}
}
\description{
\code{to_graphNEL} converts an object to a \code{\link[graph]{graphNEL}} object.
At a basic level, simply converts all associated \code{\link{edgeList}} objects to \code{graphNEL}.
}

