% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph-compatibility.R
\name{to_igraph}
\alias{to_igraph}
\title{Conversion to igraph object}
\usage{
to_igraph(x)
}
\arguments{
\item{x}{An object of type \code{\link{edgeList}}, \code{\link{sparsebnFit}}, or \code{\link{sparsebnPath}}.}
}
\description{
\code{igraph} converts an object to an \code{\link[igraph]{igraph}} object.
At a basic level, simply converts all associated \code{\link{edgeList}} objects to \code{igraph}.
}

