% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-compatibility.R
\name{to_network}
\alias{to_network}
\title{Conversion to network object}
\usage{
to_network(x)
}
\arguments{
\item{x}{An object of type \code{\link{edgeList}}, \code{\link{sparsebnFit}}, or \code{\link{sparsebnPath}}.}
}
\description{
\code{igraph} converts an object to a \code{\link[network]{network}} object.
At a basic level, simply converts all associated \code{\link{edgeList}} objects to \code{network}.
}

