% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-sparsebnData.R
\name{as.data.frame.sparsebnData}
\alias{as.data.frame.sparsebnData}
\title{Convert a sparsebnData object back to a data.frame}
\usage{
\method{as.data.frame}{sparsebnData}(x, ...)
}
\arguments{
\item{x}{a \code{\link{sparsebnData}} object.}

\item{...}{(optional) additional argument to \code{as.data.frame}.}
}
\description{
Convert a sparsebnData object back to a data.frame
}
