% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-edgeList.R, R/s3-generics.R
\docType{class}
\name{edgeList}
\alias{as.edgeList}
\alias{edgeList}
\alias{is.edgeList}
\title{edgeList class}
\usage{
is.edgeList(x)

edgeList(x)

as.edgeList(x)
}
\arguments{
\item{x}{A list containing parents for each node in a graph. The length of this list
should be the same as the number of nodes in the graph.}
}
\description{
Convenience wrapper class for a (column-major) edge list. Each component of the list
corresponds to a node, and each component is an integer vector whose components are the parents
of this node in the graph.
}
\details{
Also inherits from \code{\link{list}}.
}
\section{Methods}{

\code{\link{get.adjacency.matrix}},
\code{\link{num.nodes}}, \code{\link{num.edges}}
}

