% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-generics.R
\name{estimate.covariance}
\alias{estimate.covariance}
\alias{estimate.precision}
\title{estimate.covariance}
\usage{
estimate.covariance(fit, data, ...)

estimate.precision(x, ...)
}
\value{
\code{matrix}
}
\description{
Estimates the covariance (or concentration) matrix implied by the associated graph object.
}

