% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-generics.R, R/sparsebnUtils-covariance.R
\name{estimate.covariance}
\alias{estimate.covariance}
\alias{estimate.precision}
\alias{get.covariance}
\alias{get.precision}
\title{Covariance estimation}
\usage{
estimate.covariance(data, ...)

estimate.precision(data, ...)

get.covariance(x, data, ...)

get.precision(coefs, vars)
}
\arguments{
\item{data}{data as \code{\link{sparsebnData}} object.}

\item{...}{(optional) additional parameters}

\item{coefs}{coefficients of DAG.}

\item{vars}{conditional variances of DAG.}

\item{fit}{fitted \code{\link{sparsebnFit}} or \code{\link{sparsebnPath}} object.}
}
\value{
Covariance or precision (inverse covariance) matrix as \code{\link[Matrix]{Matrix}}.
}
\description{
Methods for inferring (i) Covariance matrices and (ii) Precision matrices for continuous,
Gaussian data.
}
\details{
For Gaussian data, the precision matrix corresponds to an undirected graphical model for the
distribution. This undirected graph can be tied to the corresponding directed graphical model;
see Sections 2.1 and 2.2 (equation (6)) of Aragam and Zhou (2015) for more details.
}

