% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-generics.R
\name{get.covariance}
\alias{get.covariance}
\alias{get.precision}
\title{Covariance and precision matrices}
\usage{
get.covariance(x, data, ...)

get.precision(x, data, ...)
}
\arguments{
\item{x}{fitted \code{\link{sparsebnFit}} or \code{\link{sparsebnPath}} object.}

\item{data}{data as \code{\link{sparsebnData}} object.}

\item{...}{(optional) additional parameters to \code{\link[sparsebn]{estimate.dag}}}
}
\value{
Covariance (or precision) matrix as \code{\link[Matrix]{Matrix}} object.
}
\description{
Methods for computing covariance and precision matrices given an estimated directed graph.
}
\details{
For Gaussian data, the precision matrix corresponds to an undirected graphical model for the
distribution. This undirected graph can be tied to the corresponding directed graphical model;
see Sections 2.1 and 2.2 (equation (6)) of Aragam and Zhou (2015) for more details.
}

