% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-sparsebnPath.R
\name{get.solution}
\alias{get.solution}
\title{Select solutions from a solution path}
\usage{
get.solution(x, edges, lambda, index)
}
\arguments{
\item{x}{a \code{\link{sparsebnPath}} object.}

\item{edges}{number of edges to search for.}

\item{lambda}{value of regularization parameter to search for.}

\item{index}{integer index to select.}
}
\description{
Choose solutions from a solution path based on number of edges, value of
regularization parameter lambda, or index.
}
\details{
If there is more than one match (for example, by number of edges), then
the first such estimate is returned. Note that \code{get.solution(x, index = j)}
is equivalent to (but slightly slower than) \code{x[[j]]}.
}

