% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-sparsebnPath.R
\name{get.solution}
\alias{get.solution}
\title{Select solutions from a solution path}
\usage{
get.solution(x, edges, lambda, index)
}
\description{
Choose solutions from a solution path based on number of edges, value of
regulariation parameter lambda, or index.
}
\details{
If there is more than one match (for example, by number of edges), then
the first such estimate is returned. Note that \code{get.solution(x, index = j)}
is equivalent to (but slightly slower than) \code{x[[j]]}.
}

