% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-edgeList.R
\name{plot.edgeList}
\alias{plot.edgeList}
\title{Plot a fitted Bayesian network object}
\usage{
\method{plot}{edgeList}(x, ...)
}
\arguments{
\item{x}{fitted object to plot.}

\item{...}{(optional) additional arguments to plotting mechanism.}
}
\description{
Plots the graph object associated with the output of a learning algorithm.
}
\details{
\code{plot.sparsebnFit} uses some default settings to make large graphs
easier to interpret, but these settings can be over-ridden.
}
\seealso{
\code{\link{setPlotPackage}}, \code{\link{getPlotPackage}}
}
