% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-sparsebnFit.R, R/s3-sparsebnPath.R
\name{plot.sparsebnFit}
\alias{plot.sparsebnFit}
\alias{plot.sparsebnPath}
\title{Plot a fitted Bayesian network object}
\usage{
\method{plot}{sparsebnFit}(x, ...)

\method{plot}{sparsebnPath}(x, ...)
}
\arguments{
\item{x}{fitted object (\code{\link{sparsebnFit}} or \code{\link{sparsebnPath}}) to plot.}

\item{...}{(optional) additional arguments to plotting mechanism.}
}
\description{
Plot plot plot
}

