% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsebnUtils-generate.R
\name{random.dag}
\alias{random.dag}
\title{Generate a random DAG with fixed number of edges}
\usage{
random.dag(nnode, nedge, call = NULL)
}
\arguments{
\item{nnode}{Number of nodes in the DAG}

\item{nedge}{Number of edges in the DAG}

\item{call}{Optional function to be used as a random number generator}
}
\description{
Generate a random DAG with fixed number of edges
}

