% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsebnUtils-show.R
\name{show.parents}
\alias{show.parents}
\title{Inspect subgraph}
\usage{
show.parents(x, nodes)
}
\arguments{
\item{x}{\code{\link{sparsebnFit}} object.}

\item{nodes}{\code{character} vector containing names of nodes to show.}
}
\description{
Print out the edge list corresponding to a subset of nodes in a graph. Useful
for inspecting particular nodes of interest in a large graph. Out is indexed
by children, with the parents of each node listed to the right of each child.
}
\details{
Uses partial matching, duplicates are OK and will be duplicated in output.
}
