% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-generics.R, R/sparsebnUtils-defaults.R, R/sparsebnUtils-functions.R, R/sparsebnUtils-options.R
\name{pick_family}
\alias{auto_count_levels}
\alias{auto_generate_levels}
\alias{capitalize}
\alias{check_if_complete_data}
\alias{check_if_data_matrix}
\alias{check_if_matrix}
\alias{check_list_class}
\alias{check_list_names}
\alias{check_list_numeric}
\alias{check_na}
\alias{check_null}
\alias{col_classes}
\alias{convert_factor_to_discrete}
\alias{cor_vector}
\alias{cor_vector_ivn}
\alias{count_nas}
\alias{default_alpha}
\alias{default_max_iters}
\alias{list_classes}
\alias{pick_family}
\alias{reIndexC}
\alias{reIndexR}
\alias{recode_levels}
\alias{sparsebn-functions}
\alias{zero_threshold}
\title{Utility functions}
\usage{
pick_family(x)

reIndexC(x)

reIndexR(x)

default_max_iters(numnode)

default_alpha()

check_if_matrix(m)

check_if_data_matrix(df)

check_if_complete_data(df)

check_null(x)

check_na(x)

count_nas(df)

list_classes(li)

auto_generate_levels(df)

auto_count_levels(df)

check_list_class(li, check.class)

check_list_numeric(li)

check_list_names(li, check.names)

col_classes(X)

capitalize(string)

recode_levels(x)

convert_factor_to_discrete(x)

cor_vector(data)

cor_vector_ivn(data, ivn)

zero_threshold()
}
\arguments{
\item{x}{a compatible object.}

\item{numnode}{\code{integer} number of nodes.}

\item{m}{a \code{matrix}.}

\item{df}{a \code{data.frame}.}

\item{li}{a \code{list}.}

\item{check.class}{\code{character} class name to compare against.}

\item{check.names}{\code{character} names to compare against.}

\item{X}{a matrix.}

\item{string}{a \code{character} string.}

\item{data}{a \code{data.frame}.}

\item{ivn}{list of interventions (see \code{\link{sparsebnData}}).}
}
\description{
Various utility functions for packages in the \code{sparsebn} family
}

