% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsebnUtils-messages.R
\name{sparsebn-messages}
\alias{alg_input_data_frame}
\alias{dag_summary}
\alias{data_not_numeric}
\alias{empty_dag_summary}
\alias{feature_not_supported}
\alias{global_coerce_warning}
\alias{has_missing_values}
\alias{input_not_sparsebnData}
\alias{invalid_class}
\alias{invalid_pkg_specification}
\alias{pkg_not_installed}
\alias{sparsebn-messages}
\title{Messages}
\usage{
input_not_sparsebnData(data)

alg_input_data_frame()

has_missing_values(count)

invalid_pkg_specification()

pkg_not_installed(pkg)

global_coerce_warning(pkg)

feature_not_supported(feature)

invalid_class(actual, expected)

dag_summary(nnode, nedge)

empty_dag_summary(nnode)

data_not_numeric(indices)
}
\arguments{
\item{data}{data object.}

\item{count}{number of missing values.}

\item{pkg}{package name.}

\item{feature}{feature name.}

\item{actual}{class input by user.}

\item{expected}{class input expected by function.}

\item{nnode}{number of nodes in a DAG.}

\item{nedge}{number of edges in a DAG.}

\item{indices}{invalid indices}
}
\description{
Warning and error messages for use in the \code{sparsebn} family
}

