% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-compatibility.R
\name{to_graphNEL}
\alias{to_graphNEL}
\title{Conversion to graphNEL object}
\usage{
to_graphNEL(x)
}
\arguments{
\item{x}{An object of type \code{\link{sparsebnPath}}, \code{\link{sparsebnFit}}, \code{\link{edgeList}},
\code{\link[igraph]{igraph}}, or \code{\link[network]{network}}.}
}
\description{
\code{to_graphNEL} converts an object to a \code{\link[graph]{graphNEL}} object. Works on both fitted
objects and graphs themselves. In the first case, every underlying 'edges' component is converted to
\code{\link[graph]{graphNEL}}. In the second, the conversion applies directly to the object.
}

