% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph-compatibility.R
\name{to_igraph}
\alias{to_igraph}
\title{Conversion to igraph object}
\usage{
to_igraph(x)
}
\arguments{
\item{x}{An object of type \code{\link{sparsebnPath}}, \code{\link{sparsebnFit}}, \code{\link{edgeList}},
\code{\link[graph]{graphNEL-class}}, or \code{\link[network]{network}}.}
}
\description{
\code{to_igraph} converts an object to a \code{\link[igraph]{igraph}} object. Works on both fitted
objects and graphs themselves. In the first case, every underlying 'edges' component is converted to
\code{\link[igraph]{igraph}}. In the second, the conversion applies directly to the object.
}

