% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-compatibility.R
\name{to_network}
\alias{to_network}
\title{Conversion to network object}
\usage{
to_network(x)
}
\arguments{
\item{x}{An object of type \code{\link{sparsebnPath}}, \code{\link{sparsebnFit}}, \code{\link{edgeList}},
\code{\link[graph]{graphNEL-class}}, or \code{\link[igraph]{igraph}}}
}
\description{
\code{to_network} converts an object to a \code{\link[network]{network}} object. Works on both fitted
objects and graphs themselves. In the first case, every underlying 'edges' component is converted to
\code{\link[network]{network}}. In the second, the conversion applies directly to the object.
}
